requirements.txt
# Python 3.9.10

db-sqlite3
Pillow
opencv-contrib-python
cmake
mediapipe
yagmail
cx_Freeze
pandas
numpy
tensorflow
smtplib